changed user delete url from a GET mapping to a DELETE and added further user authurisation to delete a user

+++ src/main/java/uk/ac/cf/cs/ons/skillsdb/skillsdb/users/UserController.java
@@ -81,11 +81,21 @@ public class UserController {
 
     }
 
-    @RequestMapping("deleteUser")
+    @DeleteMapping("deleteUser")
     public String deleteUser(@RequestParam("id") long id){
-        service.deleteUserById(id);
-        asService.deleteAssociatedSkillsByUserId(id);
-        return "index";
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User requesteduser = service.findUserByIndex(id).get();
+        String loggedinuser = authentication.getName();
+
+        if (requesteduser != null && requesteduser.getUsername().equals(loggedinuser)) {
+            service.deleteUserById(id);
+            asService.deleteAssociatedSkillsByUserId(id);
+            return "index";
+        }
+        else {
+            return "404";
+        }
 
 
     }
